Refactor project files, Add tests for services

+++ TodoApp.Domain/Services/TodoListService.cs
@@ -1,10 +1,9 @@
 ﻿using TodoApp.Domain.Entities;
-using TodoApp.Domain.Models;
 using TodoApp.Domain.Repositories;
 
 namespace TodoApp.Domain.Services
 {
-    public class TodoListService
+    public class TodoListService : ITodoListService
     {
         private readonly ITodoListRepository repository;
 
@@ -17,16 +16,28 @@ namespace TodoApp.Domain.Services
 
         public TodoList Add(TodoList todoList)
         {
+            if (this.repository.TodoLists.Where(t => t.Id == todoList.Id).Any())
+            {
+                throw new InvalidOperationException("TodoList already exists");
+            }
             return this.repository.Add(todoList);
         }
 
         public TodoList Update(TodoList todoList)
         {
+            if (!this.repository.TodoLists.Where(t => t.Id == todoList.Id).Any())
+            {
+                throw new InvalidOperationException("TodoList does not exist");
+            }
             return this.repository.Update(todoList);
         }
 
         public TodoList Delete(TodoList todoList)
         {
+            if (!this.repository.TodoLists.Where(t => t.Id == todoList.Id).Any())
+            {
+                throw new InvalidOperationException("TodoList does not exist");
+            }
             return this.repository.Delete(todoList);
         }
     }
